requirements.txt
django
djangorestframework
djangorestframework-simplejwt
coreapi>=2.3.3
drf-yasg>=1.21.5
channels
channels_postgres